[ML] using feature reset API for integration test clean up: external multi-node
