+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/CategorizationIT.java
@@ -111,7 +111,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
     public void testBasicCategorization() throws Exception {
         Job.Builder job = newJobBuilder("categorization", Collections.emptyList(), false);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -119,7 +118,7 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
         datafeedConfig.setIndices(Collections.singletonList(DATA_INDEX));
         DatafeedConfig datafeed = datafeedConfig.build();
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         startDatafeed(datafeedId, 0, nowMillis);
         waitUntilJobIsClosed(job.getId());
@@ -169,7 +168,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
     public void testPerPartitionCategorization() throws Exception {
         Job.Builder job = newJobBuilder("per-partition-categorization", Collections.emptyList(), true);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -177,7 +175,7 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
         datafeedConfig.setIndices(Collections.singletonList(DATA_INDEX));
         DatafeedConfig datafeed = datafeedConfig.build();
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         startDatafeed(datafeedId, 0, nowMillis);
         waitUntilJobIsClosed(job.getId());
@@ -248,7 +246,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
     public void testCategorizationWithFilters() throws Exception {
         Job.Builder job = newJobBuilder("categorization-with-filters", Collections.singletonList("\\[.*\\]"), false);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -256,7 +253,7 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
         datafeedConfig.setIndices(Collections.singletonList(DATA_INDEX));
         DatafeedConfig datafeed = datafeedConfig.build();
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         startDatafeed(datafeedId, 0, nowMillis);
         waitUntilJobIsClosed(job.getId());
@@ -278,7 +275,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
     public void testCategorizationStatePersistedOnSwitchToRealtime() throws Exception {
         Job.Builder job = newJobBuilder("categorization-swtich-to-realtime", Collections.emptyList(), false);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -286,7 +282,7 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
         datafeedConfig.setIndices(Collections.singletonList(DATA_INDEX));
         DatafeedConfig datafeed = datafeedConfig.build();
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         startDatafeed(datafeedId, 0, null);
 
@@ -355,7 +351,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
         String jobId = "categorization-performance";
         Job.Builder job = newJobBuilder(jobId, Collections.emptyList(), false);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -382,7 +377,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
 
         String jobId = "categorization-stop-on-warn";
         Job.Builder job = newJobBuilder(jobId, Collections.emptyList(), true);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -526,7 +520,6 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         assertThat(bulkResponse.hasFailures(), is(false));
 
         Job.Builder job = newJobBuilder("categorization-with-preferred-categories", Collections.emptyList(), false);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
 
@@ -534,7 +527,7 @@ public class CategorizationIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
         datafeedConfig.setIndices(Collections.singletonList(index));
         DatafeedConfig datafeed = datafeedConfig.build();
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         startDatafeed(datafeedId, 0, nowMillis + 1);
         waitUntilJobIsClosed(job.getId());
