+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/DatafeedJobsIT.java
@@ -98,7 +98,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         indexDocs(logger, "data-2", numDocs2, oneWeekAgo, now);
 
         Job.Builder job = createScheduledJob("lookback-job");
-        registerJob(job);
         PutJobAction.Response putJobResponse = putJob(job);
         assertThat(putJobResponse.getResponse().getJobVersion(), equalTo(Version.CURRENT));
         openJob(job.getId());
@@ -107,7 +106,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         // Having a pattern with missing indices is acceptable
         List<String> indices = Arrays.asList("data-*", "missing-*");
         DatafeedConfig datafeedConfig = createDatafeed(job.getId() + "-datafeed", job.getId(), indices);
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
 
         startDatafeed(datafeedConfig.getId(), 0L, now);
@@ -145,7 +143,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         client().admin().cluster().prepareHealth("datafeed_data_stream").setWaitForYellowStatus().get();
 
         Job.Builder job = createScheduledJob("lookback-data-stream-job");
-        registerJob(job);
         PutJobAction.Response putJobResponse = putJob(job);
         assertThat(putJobResponse.getResponse().getJobVersion(), equalTo(Version.CURRENT));
         openJob(job.getId());
@@ -154,7 +151,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig datafeedConfig = createDatafeed(job.getId() + "-datafeed",
             job.getId(),
             Collections.singletonList("datafeed_data_stream"));
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
 
         startDatafeed(datafeedConfig.getId(), 0L, now);
@@ -195,7 +191,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         jobBuilder.setAnalysisConfig(analysisConfig);
         jobBuilder.setDataDescription(dataDescription);
 
-        registerJob(jobBuilder);
         putJob(jobBuilder);
         openJob(jobBuilder.getId());
         assertBusy(() -> assertEquals(getJobStats(jobBuilder.getId()).get(0).getState(), JobState.OPENED));
@@ -214,7 +209,7 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
 
         DatafeedConfig datafeedConfig = dfBuilder.build();
 
-        registerDatafeed(datafeedConfig);
+
         putDatafeed(datafeedConfig);
 
         startDatafeed(datafeedConfig.getId(), 0L, now);
@@ -246,13 +241,12 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         String datafeedId = "lookback-datafeed-datafeed-recreated";
         DatafeedConfig datafeedConfig = createDatafeed(datafeedId, job.getId(), Collections.singletonList("data"));
 
-        registerJob(job);
         putJob(job);
 
         CheckedRunnable<Exception> openAndRunJob = () -> {
             openJob(job.getId());
             assertBusy(() -> assertEquals(getJobStats(job.getId()).get(0).getState(), JobState.OPENED));
-            registerDatafeed(datafeedConfig);
+
             putDatafeed(datafeedConfig);
             // Datafeed did not do anything yet, hence search_count is equal to 0.
             assertDatafeedStats(datafeedId, DatafeedState.STOPPED, job.getId(), equalTo(0L));
@@ -279,12 +273,10 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         indexDocs(logger, "data", numDocs, now.minus(Duration.ofDays(14)).toEpochMilli(), now.toEpochMilli());
 
         Job.Builder job = createScheduledJob("lookback-job-query-delay-updated");
-        registerJob(job);
         putJob(job);
 
         String datafeedId = "lookback-datafeed-query-delay-updated";
         DatafeedConfig datafeedConfig = createDatafeed(datafeedId, job.getId(), Collections.singletonList("data"));
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
 
         openJob(job.getId());
@@ -327,7 +319,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
 
         String scrollJobId = "stop-restart-scroll";
         Job.Builder scrollJob = createScheduledJob(scrollJobId);
-        registerJob(scrollJob);
         putJob(scrollJob);
         openJob(scrollJobId);
         assertBusy(() -> assertEquals(getJobStats(scrollJobId).get(0).getState(), JobState.OPENED));
@@ -335,7 +326,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         DatafeedConfig datafeedConfig = createDatafeedBuilder(scrollJobId+ "-datafeed", scrollJobId, Collections.singletonList(indexName))
             .setChunkingConfig(ChunkingConfig.newManual(new TimeValue(1, TimeUnit.SECONDS)))
             .build();
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
         startDatafeed(datafeedConfig.getId(), 0L, null);
 
@@ -365,7 +355,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         compositeJob.setAnalysisConfig(
             new AnalysisConfig.Builder(compositeJob.getAnalysisConfig()).setSummaryCountFieldName("doc_count")
         );
-        registerJob(compositeJob);
         putJob(compositeJob);
         openJob(compositeJobId);
         assertBusy(() -> assertEquals(getJobStats(compositeJobId).get(0).getState(), JobState.OPENED));
@@ -390,7 +379,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
             // Start off chunking at an hour so that it runs more slowly and the test has time to stop it in the middle of processing
             .setChunkingConfig(ChunkingConfig.newManual(TimeValue.timeValueHours(1)))
             .build();
-        registerDatafeed(compositeDatafeedConfig);
         putDatafeed(compositeDatafeedConfig);
         startDatafeed(compositeDatafeedConfig.getId(), 0L, null);
 
@@ -617,7 +605,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         indexDocs(logger, "data", numDocs, twoWeeksAgo, oneWeekAgo);
 
         Job.Builder job = createScheduledJob("lookback-job-stopped-then-killed");
-        registerJob(job);
         PutJobAction.Response putJobResponse = putJob(job);
         assertThat(putJobResponse.getResponse().getJobVersion(), equalTo(Version.CURRENT));
         openJob(job.getId());
@@ -628,7 +615,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         // Use lots of chunks so we have time to stop the lookback before it completes
         datafeedConfigBuilder.setChunkingConfig(ChunkingConfig.newManual(new TimeValue(1, TimeUnit.SECONDS)));
         DatafeedConfig datafeedConfig = datafeedConfigBuilder.build();
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
         startDatafeed(datafeedConfig.getId(), 0L, now);
         assertBusy(() -> {
@@ -671,7 +657,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
         }
 
         Job.Builder job = createScheduledJob(jobId);
-        registerJob(job);
         putJob(job);
         openJob(job.getId());
         assertBusy(() -> assertEquals(getJobStats(job.getId()).get(0).getState(), JobState.OPENED));
@@ -682,7 +667,6 @@ public class DatafeedJobsIT extends MlNativeAutodetectIntegTestCase {
             datafeedConfigBuilder.setMaxEmptySearches(maxEmptySearches);
         }
         DatafeedConfig datafeedConfig = datafeedConfigBuilder.build();
-        registerDatafeed(datafeedConfig);
         putDatafeed(datafeedConfig);
         startDatafeed(datafeedConfig.getId(), 0L, null);
         assertBusy(() -> {
