+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/DeleteExpiredDataIT.java
@@ -157,17 +157,18 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
             bulkRequestBuilder.add(indexRequest);
         }
         ActionFuture<BulkResponse> indexUnusedStateDocsResponse = bulkRequestBuilder.execute();
+        List<Job.Builder> jobs = new ArrayList<>();
 
         // These jobs don't thin out model state; ModelSnapshotRetentionIT tests that
-        registerJob(newJobBuilder("no-retention")
+        jobs.add(newJobBuilder("no-retention")
             .setResultsRetentionDays(null).setModelSnapshotRetentionDays(1000L).setDailyModelSnapshotRetentionAfterDays(1000L));
-        registerJob(newJobBuilder("results-retention")
+        jobs.add(newJobBuilder("results-retention")
             .setResultsRetentionDays(1L).setModelSnapshotRetentionDays(1000L).setDailyModelSnapshotRetentionAfterDays(1000L));
-        registerJob(newJobBuilder("snapshots-retention")
+        jobs.add(newJobBuilder("snapshots-retention")
             .setResultsRetentionDays(null).setModelSnapshotRetentionDays(2L).setDailyModelSnapshotRetentionAfterDays(2L));
-        registerJob(newJobBuilder("snapshots-retention-with-retain")
+        jobs.add(newJobBuilder("snapshots-retention-with-retain")
             .setResultsRetentionDays(null).setModelSnapshotRetentionDays(2L).setDailyModelSnapshotRetentionAfterDays(2L));
-        registerJob(newJobBuilder("results-and-snapshots-retention")
+        jobs.add(newJobBuilder("results-and-snapshots-retention")
             .setResultsRetentionDays(1L).setModelSnapshotRetentionDays(2L).setDailyModelSnapshotRetentionAfterDays(2L));
 
         List<String> shortExpiryForecastIds = new ArrayList<>();
@@ -176,14 +177,14 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
         long oneDayAgo = now - TimeValue.timeValueHours(48).getMillis() - 1;
 
         // Start all jobs
-        for (Job.Builder job : getJobs()) {
+        for (Job.Builder job : jobs) {
             putJob(job);
 
             String datafeedId = job.getId() + "-feed";
             DatafeedConfig.Builder datafeedConfig = new DatafeedConfig.Builder(datafeedId, job.getId());
             datafeedConfig.setIndices(Collections.singletonList(DATA_INDEX));
             DatafeedConfig datafeed = datafeedConfig.build();
-            registerDatafeed(datafeed);
+
             putDatafeed(datafeed);
 
             // Run up to a day ago
@@ -192,11 +193,11 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
         }
 
         // Now let's wait for all jobs to be closed
-        for (Job.Builder job : getJobs()) {
+        for (Job.Builder job : jobs) {
             waitUntilJobIsClosed(job.getId());
         }
 
-        for (Job.Builder job : getJobs()) {
+        for (Job.Builder job : jobs) {
             assertThat(getBuckets(job.getId()).size(), is(greaterThanOrEqualTo(47)));
             assertThat(getRecords(job.getId()).size(), equalTo(2));
             List<ModelSnapshot> modelSnapshots = getModelSnapshots(job.getId());
@@ -231,7 +232,7 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
         long before = System.currentTimeMillis() / 1000;
         assertBusy(() -> assertNotEquals(before, System.currentTimeMillis() / 1000), 1, TimeUnit.SECONDS);
 
-        for (Job.Builder job : getJobs()) {
+        for (Job.Builder job : jobs) {
             // Run up to now
             startDatafeed(job.getId() + "-feed", 0, now);
             waitUntilJobIsClosed(job.getId());
@@ -254,7 +255,7 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
 
         // Verify forecasts were created
         List<ForecastRequestStats> forecastStats = getForecastStats();
-        assertThat(forecastStats.size(), equalTo(getJobs().size() * 3));
+        assertThat(forecastStats.size(), equalTo(jobs.size() * 3));
         for (ForecastRequestStats forecastStat : forecastStats) {
             assertThat(countForecastDocs(forecastStat.getJobId(), forecastStat.getForecastId()), equalTo(forecastStat.getRecordCount()));
         }
@@ -303,11 +304,11 @@ public class DeleteExpiredDataIT extends MlNativeAutodetectIntegTestCase {
 
         // Verify short expiry forecasts were deleted only
         forecastStats = getForecastStats();
-        assertThat(forecastStats.size(), equalTo(getJobs().size() * 2));
+        assertThat(forecastStats.size(), equalTo(jobs.size() * 2));
         for (ForecastRequestStats forecastStat : forecastStats) {
             assertThat(countForecastDocs(forecastStat.getJobId(), forecastStat.getForecastId()), equalTo(forecastStat.getRecordCount()));
         }
-        for (Job.Builder job : getJobs()) {
+        for (Job.Builder job : jobs) {
             for (String forecastId : shortExpiryForecastIds) {
                 assertThat(countForecastDocs(job.getId(), forecastId), equalTo(0L));
             }
