+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/InterimResultsIT.java
@@ -50,7 +50,7 @@ public class InterimResultsIT extends MlNativeAutodetectIntegTestCase {
         job.setAnalysisConfig(analysisConfig);
         job.setDataDescription(dataDescription);
 
-        registerJob(job);
+
         putJob(job);
         openJob(job.getId());
 
@@ -118,7 +118,7 @@ public class InterimResultsIT extends MlNativeAutodetectIntegTestCase {
         job.setAnalysisConfig(analysisConfig);
         job.setDataDescription(dataDescription);
 
-        registerJob(job);
+
         putJob(job);
         openJob(job.getId());
