+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/MlDailyMaintenanceServiceIT.java
@@ -112,7 +112,7 @@ public class MlDailyMaintenanceServiceIT extends MlNativeAutodetectIntegTestCase
                     new DataDescription.Builder()
                         .setTimeFormat("epoch"));
 
-        registerJob(job);
+
         putJob(job);
     }
