+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/MlNativeAutodetectIntegTestCase.java
@@ -6,6 +6,8 @@
  */
 package org.elasticsearch.xpack.ml.integration;
 
+import org.elasticsearch.action.admin.cluster.snapshots.features.ResetFeatureStateAction;
+import org.elasticsearch.action.admin.cluster.snapshots.features.ResetFeatureStateRequest;
 import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.support.IndicesOptions;
@@ -86,69 +88,6 @@ import static org.hamcrest.Matchers.notNullValue;
  * Base class of ML integration tests that use a native autodetect process
  */
 abstract class MlNativeAutodetectIntegTestCase extends MlNativeIntegTestCase {
-
-    private final List<Job.Builder> jobs = new ArrayList<>();
-    private final List<DatafeedConfig> datafeeds = new ArrayList<>();
-
-    @Override
-    protected void cleanUpResources() {
-        cleanUpDatafeeds();
-        cleanUpJobs();
-    }
-
-    private void cleanUpDatafeeds() {
-        for (DatafeedConfig datafeed : datafeeds) {
-            cleanupDatafeed(datafeed.getId());
-        }
-    }
-
-    void cleanupDatafeed(String datafeedId) {
-        try {
-            stopDatafeed(datafeedId);
-        } catch (Exception e) {
-            // ignore
-        }
-        try {
-            deleteDatafeed(datafeedId);
-        } catch (Exception e) {
-            // ignore
-        }
-    }
-
-    private void cleanUpJobs() {
-        for (Job.Builder job : jobs) {
-            cleanupJob(job.getId());
-        }
-    }
-
-    void cleanupJob(String jobId) {
-        try {
-            closeJob(jobId);
-        } catch (Exception e) {
-            // ignore
-        }
-        try {
-            deleteJob(jobId);
-        } catch (Exception e) {
-            // ignore
-        }
-    }
-
-    protected void registerJob(Job.Builder job) {
-        if (jobs.add(job) == false) {
-            throw new IllegalArgumentException("job [" + job.getId() + "] is already registered");
-        }
-    }
-
-    protected void registerDatafeed(DatafeedConfig datafeed) {
-        if (datafeeds.add(datafeed) == false) {
-            throw new IllegalArgumentException("datafeed [" + datafeed.getId() + "] is already registered");
-        }
-    }
-
-    protected List<Job.Builder> getJobs() {
-        return jobs;
-    }
 
     protected PutJobAction.Response putJob(Job.Builder job) {
         PutJobAction.Request request = new PutJobAction.Request(job);
