+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/MlNativeDataFrameAnalyticsIntegTestCase.java
@@ -6,7 +6,6 @@
  */
 package org.elasticsearch.xpack.ml.integration;
 
-import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.elasticsearch.action.admin.indices.get.GetIndexAction;
 import org.elasticsearch.action.admin.indices.get.GetIndexRequest;
 import org.elasticsearch.action.admin.indices.refresh.RefreshAction;
@@ -16,7 +15,6 @@ import org.elasticsearch.action.search.SearchAction;
 import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.Strings;
@@ -28,7 +26,6 @@ import org.elasticsearch.search.SearchHit;
 import org.elasticsearch.search.sort.SortOrder;
 import org.elasticsearch.tasks.TaskInfo;
 import org.elasticsearch.xpack.core.ml.MlTasks;
-import org.elasticsearch.xpack.core.ml.action.DeleteDataFrameAnalyticsAction;
 import org.elasticsearch.xpack.core.ml.action.EvaluateDataFrameAction;
 import org.elasticsearch.xpack.core.ml.action.ExplainDataFrameAnalyticsAction;
 import org.elasticsearch.xpack.core.ml.action.PreviewDataFrameAnalyticsAction;
@@ -56,7 +53,6 @@ import org.elasticsearch.xpack.ml.dataframe.StoredProgress;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -84,44 +80,7 @@ import static org.hamcrest.Matchers.nullValue;
  */
 abstract class MlNativeDataFrameAnalyticsIntegTestCase extends MlNativeIntegTestCase {
 
-    private final List<DataFrameAnalyticsConfig> analytics = new ArrayList<>();
-
-    @Override
-    protected void cleanUpResources() {
-        cleanUpAnalytics();
-    }
-
-    private void cleanUpAnalytics() {
-        stopAnalyticsAndForceStopOnError();
-
-        for (DataFrameAnalyticsConfig config : analytics) {
-            try {
-                assertThat(deleteAnalytics(config.getId()).isAcknowledged(), is(true));
-            } catch (Exception e) {
-                // just log and ignore
-                logger.error(new ParameterizedMessage("[{}] Could not clean up analytics job config", config.getId()), e);
-            }
-        }
-    }
-
-    private void stopAnalyticsAndForceStopOnError() {
-        try {
-            assertThat(stopAnalytics("*").isStopped(), is(true));
-        } catch (Exception e) {
-            logger.error("Failed to stop data frame analytics jobs; trying force", e);
-            try {
-                assertThat(forceStopAnalytics("*").isStopped(), is(true));
-            } catch (Exception e2) {
-                logger.error("Force-stopping data frame analytics jobs failed", e2);
-            }
-            throw new RuntimeException("Had to resort to force-stopping jobs, something went wrong?", e);
-        }
-    }
-
     protected PutDataFrameAnalyticsAction.Response putAnalytics(DataFrameAnalyticsConfig config) {
-        if (analytics.add(config) == false) {
-            throw new IllegalArgumentException("analytics config [" + config.getId() + "] is already registered");
-        }
         PutDataFrameAnalyticsAction.Request request = new PutDataFrameAnalyticsAction.Request(config);
         return client().execute(PutDataFrameAnalyticsAction.INSTANCE, request).actionGet();
     }
@@ -129,11 +88,6 @@ abstract class MlNativeDataFrameAnalyticsIntegTestCase extends MlNativeIntegTest
     protected PutDataFrameAnalyticsAction.Response updateAnalytics(DataFrameAnalyticsConfigUpdate update) {
         UpdateDataFrameAnalyticsAction.Request request = new UpdateDataFrameAnalyticsAction.Request(update);
         return client().execute(UpdateDataFrameAnalyticsAction.INSTANCE, request).actionGet();
-    }
-
-    protected AcknowledgedResponse deleteAnalytics(String id) {
-        DeleteDataFrameAnalyticsAction.Request request = new DeleteDataFrameAnalyticsAction.Request(id);
-        return client().execute(DeleteDataFrameAnalyticsAction.INSTANCE, request).actionGet();
     }
 
     protected NodeAcknowledgedResponse startAnalytics(String id) {
