+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/ModelPlotsIT.java
@@ -74,11 +74,11 @@ public class ModelPlotsIT extends MlNativeAutodetectIntegTestCase {
     public void testPartitionFieldWithoutTerms() throws Exception {
         Job.Builder job = jobWithPartitionUser("model-plots-it-test-partition-field-without-terms");
         job.setModelPlotConfig(new ModelPlotConfig());
-        registerJob(job);
+
         putJob(job);
         String datafeedId = job.getId() + "-feed";
         DatafeedConfig datafeed = newDatafeed(datafeedId, job.getId());
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         openJob(job.getId());
         startDatafeed(datafeedId, 0, System.currentTimeMillis());
@@ -96,11 +96,11 @@ public class ModelPlotsIT extends MlNativeAutodetectIntegTestCase {
     public void testPartitionFieldWithTerms() throws Exception {
         Job.Builder job = jobWithPartitionUser("model-plots-it-test-partition-field-with-terms");
         job.setModelPlotConfig(new ModelPlotConfig(true, "user_2,user_3", false));
-        registerJob(job);
+
         putJob(job);
         String datafeedId = job.getId() + "-feed";
         DatafeedConfig datafeed = newDatafeed(datafeedId, job.getId());
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         openJob(job.getId());
         startDatafeed(datafeedId, 0, System.currentTimeMillis());
@@ -118,11 +118,11 @@ public class ModelPlotsIT extends MlNativeAutodetectIntegTestCase {
     public void testByFieldWithTerms() throws Exception {
         Job.Builder job = jobWithByUser("model-plots-it-test-by-field-with-terms");
         job.setModelPlotConfig(new ModelPlotConfig(true, "user_2,user_3", false));
-        registerJob(job);
+
         putJob(job);
         String datafeedId = job.getId() + "-feed";
         DatafeedConfig datafeed = newDatafeed(datafeedId, job.getId());
-        registerDatafeed(datafeed);
+
         putDatafeed(datafeed);
         openJob(job.getId());
         startDatafeed(datafeedId, 0, System.currentTimeMillis());
