+++ x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/PersistJobIT.java
@@ -216,7 +216,7 @@ public class PersistJobIT extends MlNativeAutodetectIntegTestCase {
         Job.Builder job = new Job.Builder(jobId);
         job.setAnalysisConfig(analysisConfig);
         job.setDataDescription(new DataDescription.Builder());
-        registerJob(job);
+
         putJob(job);
 
         openJob(job.getId());
